Use a traditional builder pattern for creating Passes
To create a pass, first create a `PassBuilder`, then call builder methods on it, then pass it to `Frame::add_pass`. This flow seems more familiar than `start_pass/finish`.

Since the frame is no longer borrowed when building a pass, passes can now be built out-of-order (for better or worse).