Removed GpuFuture, moved `wait` method to device, use QueueSerialNumbers (now QueueProgress) directly.
GpuFuture was a useless wrapper over queue serial numbers, and wasn't actually an awaitable future. Leave the responsibility of implementing async stuff to a higher-level library.

The Context::wait_for has been moved to Device::wait, as we shouldn't need to access the context for waiting for completion of commands. It was in context before because the timelines were allocated there, and serial numbers were actually relative to a particular `Context`, potentially allowing multiple Contexts at the same time on the same device. However, this would have been incorrect because resource tracking assumes a single context and a single set of timelines. Also, Context::new takes a device by value so it prevents that.

TODO: will probably remove frame pacing from graal, and let a higher level library handle that