Make SwapchainImages move-only, and consume them in `Frame::present`.
Now swapchain images are created by `Context::acquire_next_swapchain_image`, and **must** be consumed by a call to `frame.present()`, which is then submitted to the GPU.
This removes the need to manually destroy `swapchain_image.image_id`, which was easily forgotten.

Also: remove the `name` parameter on `Frame::present` for consistency with other pass methods.

TODO Add a way to specify a name for a present pass. Fairly low priority.